+++ src/controller/market.ts
@@ -16,13 +16,4 @@ const createMarket = async (req: Request, res: Response) => {
   res.status(201).json({ message: 'SUCCESS' });
 };
 
-const createProduct = async (req: Request, res: Response) => {
-  const userId = req.userId;
-  const data = req.body;
-
-  await marketService.createProduct(userId!, data);
-
-  res.status(201).json({ message: 'SUCCESS' });
-};
-
-export { createMarket, createProduct };
+export { createMarket };
