+++ src/db/schema/product.ts
@@ -28,10 +28,11 @@ export const productSchema = new mongoose.Schema({
   mainCategory: { type: String, required: true },
   subCategory: { type: String, required: true },
   productInfo: { type: String, required: true },
+  noticeCheck: { type: Boolean, required: true },
   purchaseDate: { type: Date },
   price: { type: Number, required: true },
   optionType: { type: String, enum: ['single', 'group'] },
-  optionList: [optionsSchema],
+  optionsList: [optionsSchema],
   productImg: [productImgSchema],
   productDetail: [productDetailSchema],
   deliveryInfo: deliveryInfoSchema,
